Add drawer in the logged in subscription page
