+++ front/lib/amplitude/browser/index.ts
@@ -14,7 +14,7 @@ export function getBrowserClient() {
   ampli.load({
     // The environment property is a depreacted value, but still needed by the SDK. We don't use it.
     environment: "dustprod",
-    disabled: disabled,
+    disabled: false,
     client: {
       apiKey: AMPLITUDE_PUBLIC_API_KEY,
       configuration: {
